requirements.txt
pyhamcrest
filemagic
